refactor: Change AwsGovCloud to use CreateAws()
